Make the version output more flexible
This allows users to configure, what will be reported when requesting the
/version-endpoint.

In case a user has the environment variable `ACTINIA_ADDITIONAL_VERSION_INFO`
its content will be parsed as key-value-pairs (key and value seperated by
colon `:`, pairs seperated by  pipe `|`).

The content of the environment variable `ACTINIA_RUNNING_SINCE` will
also be reported, if it is set.